Protect private routes
