+++ src/_layout/Menu.tsx
@@ -1,33 +1,11 @@
-'use client';
-
-import { TbMenu2 } from 'react-icons/tb';
 import React from 'react';
-import { usePathname } from 'next/navigation';
-import Link from 'next/link';
+import { TbMenu2 } from 'react-icons/tb';
+import { MenuItem } from './MenuItem';
+import { getAuth } from '@/auth';
 
-const MenuItem: React.FC<React.PropsWithChildren<{ href: string }>> = ({
-  children,
-  href,
-}) => {
-  const pathname = usePathname();
-  const isActive = pathname === href;
+export const Menu = async () => {
+  const auth = await getAuth();
 
-  if (isActive) {
-    return (
-      <li>
-        <span className='btn-active btn-disabled'>{children}</span>
-      </li>
-    );
-  }
-
-  return (
-    <li>
-      <Link href={href}>{children}</Link>
-    </li>
-  );
-};
-
-export const Menu = () => {
   return (
     <div className='dropdown dropdown-end'>
       <div tabIndex={0} role='button' className='btn btn-square btn-ghost'>
@@ -38,8 +16,8 @@ export const Menu = () => {
         className='dropdown-content menu bg-base-100 rounded-box z-[1] w-52 p-2 shadow'
       >
         <MenuItem href='/'>Dashboard</MenuItem>
-        <MenuItem href='/share'>Share List</MenuItem>
-        <MenuItem href='/settings'>Settings</MenuItem>
+        {auth && <MenuItem href='/share'>Share List</MenuItem>}
+        {auth && <MenuItem href='/settings'>Settings</MenuItem>}
         <MenuItem href='/about'>About</MenuItem>
       </ul>
     </div>
