+++ src/app/share/page.tsx
@@ -1,8 +1,13 @@
 import React from 'react';
+import { redirect } from 'next/navigation';
 import { getAuth } from '@/auth';
 
 export default async function Share() {
   const auth = await getAuth();
+
+  if (!auth) {
+    redirect('/');
+  }
 
   return (
     <div className='mx-3 mt-3 flex-1'>
